refactor image resolver code into imageresolver package
